feat: [SCRUM-154] folder refactor
