+++ src/features/session/ui/SessionsList.jsx
